pyproject.toml
[project]
name = "datario"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandera>=0.22.1",
    "prefect==0.15.9",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.26.0",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
]
